Project structure refactoring
joined directories:
 - plugin;
 - plugin_interfaces
 - private_factory.
