--- pfactory/private_factory/internal_object_creator_private.h
@@ -1,30 +1,0 @@
-#pragma once
-#include "pfactory/plugin/object_creator.h"
-
-namespace pf
-{
-  template<typename Interface, typename Obj>
-  class InternalObjectCreatorPrivate
-      : public IObjectCreator<Interface>
-  {
-  public:
-    InternalObjectCreatorPrivate(const QString &id, const QString &interface )
-      : IObjectCreator<Interface> (id, interface )
-    {
-
-    }
-    ~InternalObjectCreatorPrivate()
-    {
-
-    }
-  public:
-
-    ///
-    /// \brief Create object
-    ///
-    Interface *create() const
-    {
-      return new Obj();
-    }
-  };
-}
