+++ pfactory/private_factory/psys/loader/plugin_loader_private.cpp
