+++ pfactory/private_factory/psys/loader/plugin_loader_private.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <QHash>
-#include "pfactory/plugin/psys/loader/iplugin_loader.h"
+#include "iplugin_loader.h"
 
 namespace psys
 {
