+++ pfactory/private_factory/template_plugin.cpp
