+++ pfactory/private_factory/template_plugin.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <QMap>
-#include "pfactory/plugin/psys/iplugin.h"
+#include "psys/iplugin.h"
 #include "sub_plugin_creator.h"
